requirements.txt
requests==2.28.1
beautifulsoup4==4.11.1
python-dotenv==1.0.0
